+++ libsolidity/formal/SymbolicVariables.cpp
@@ -27,10 +27,12 @@ using namespace dev::solidity::smt;
 
 SymbolicVariable::SymbolicVariable(
 	solidity::TypePointer _type,
+	solidity::TypePointer _originalType,
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	m_type(move(_type)),
+	m_type(_type),
+	m_originalType(_originalType),
 	m_uniqueName(move(_uniqueName)),
 	m_context(_context),
 	m_ssa(make_unique<SSAVariable>())
@@ -53,7 +55,7 @@ SymbolicVariable::SymbolicVariable(
 	solAssert(m_sort, "");
 }
 
-Expression SymbolicVariable::currentValue() const
+Expression SymbolicVariable::currentValue(solidity::TypePointer const&) const
 {
 	return valueAtIndex(m_ssa->index());
 }
@@ -95,17 +97,18 @@ SymbolicBoolVariable::SymbolicBoolVariable(
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicVariable(move(_type), move(_uniqueName), _context)
+	SymbolicVariable(_type, _type, move(_uniqueName), _context)
 {
 	solAssert(m_type->category() == solidity::Type::Category::Bool, "");
 }
 
 SymbolicIntVariable::SymbolicIntVariable(
 	solidity::TypePointer _type,
+	solidity::TypePointer _originalType,
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicVariable(move(_type), move(_uniqueName), _context)
+	SymbolicVariable(_type, _originalType, move(_uniqueName), _context)
 {
 	solAssert(isNumber(m_type->category()), "");
 }
@@ -114,16 +117,17 @@ SymbolicAddressVariable::SymbolicAddressVariable(
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicIntVariable(TypeProvider::uint(160), move(_uniqueName), _context)
+	SymbolicIntVariable(TypeProvider::uint(160), TypeProvider::uint(160), move(_uniqueName), _context)
 {
 }
 
 SymbolicFixedBytesVariable::SymbolicFixedBytesVariable(
+	solidity::TypePointer _originalType,
 	unsigned _numBytes,
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicIntVariable(TypeProvider::uint(_numBytes * 8), move(_uniqueName), _context)
+	SymbolicIntVariable(TypeProvider::uint(_numBytes * 8), _originalType, move(_uniqueName), _context)
 {
 }
 
@@ -132,7 +136,7 @@ SymbolicFunctionVariable::SymbolicFunctionVariable(
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicVariable(move(_type), move(_uniqueName), _context),
+	SymbolicVariable(_type, _type, move(_uniqueName), _context),
 	m_declaration(m_context.newVariable(currentName(), m_sort))
 {
 	solAssert(m_type->category() == solidity::Type::Category::Function, "");
@@ -171,19 +175,33 @@ SymbolicMappingVariable::SymbolicMappingVariable(
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicVariable(move(_type), move(_uniqueName), _context)
+	SymbolicVariable(_type, _type, move(_uniqueName), _context)
 {
 	solAssert(isMapping(m_type->category()), "");
 }
 
 SymbolicArrayVariable::SymbolicArrayVariable(
 	solidity::TypePointer _type,
+	solidity::TypePointer _originalType,
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicVariable(move(_type), move(_uniqueName), _context)
+	SymbolicVariable(_type, _originalType, move(_uniqueName), _context)
 {
 	solAssert(isArray(m_type->category()), "");
+}
+
+Expression SymbolicArrayVariable::currentValue(solidity::TypePointer const& _targetType) const
+{
+	if (_targetType)
+		// StringLiterals are encoded as SMT arrays in the generic case,
+		// but they can also be compared/assigned to fixed bytes, in which
+		// case they'd need to be encoded as numbers.
+		if (auto strType = dynamic_cast<StringLiteralType const*>(m_originalType))
+			if (_targetType->category() == solidity::Type::Category::FixedBytes)
+				return smt::Expression(u256(toHex(asBytes(strType->value()), HexPrefix::Add)));
+
+	return SymbolicVariable::currentValue(_targetType);
 }
 
 SymbolicEnumVariable::SymbolicEnumVariable(
@@ -191,7 +209,7 @@ SymbolicEnumVariable::SymbolicEnumVariable(
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicVariable(move(_type), move(_uniqueName), _context)
+	SymbolicVariable(_type, _type, move(_uniqueName), _context)
 {
 	solAssert(isEnum(m_type->category()), "");
 }
@@ -201,7 +219,7 @@ SymbolicTupleVariable::SymbolicTupleVariable(
 	string _uniqueName,
 	EncodingContext& _context
 ):
-	SymbolicVariable(move(_type), move(_uniqueName), _context)
+	SymbolicVariable(_type, _type, move(_uniqueName), _context)
 {
 	solAssert(isTuple(m_type->category()), "");
 }