+++ test/libsolidity/smtCheckerTests/types/string_2.sol
@@ -8,4 +8,3 @@ contract C
 }
 // ----
 // Warning: (76-91): Unused local variable.
-// Warning: (94-107): Assertion checker does not yet support the type of this literal (literal_string "Hello World").